requirements.txt
telethon==1.28.5
python-telegram-bot==20.7
python-dotenv==1.0.0
black==23.12.1
pylint==3.0.3